requirements.txt
requests
python-dotenv
ipinfo